requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asttokens=2.4.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=openblas
boost=1.74.0=py310h7c3ba0c_5
boost-cpp=1.74.0=h75c5d50_8
bottleneck=1.3.5=py310ha9d4c09_0
brotli-python=1.0.9=py310h6a678d5_7
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.08.22=h06a4308_0
cairo=1.16.0=hb05425b_5
certifi=2023.7.22=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.1.4=pyhd8ed1ab_0
cryptography=41.0.3=py310hdda0065_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.52=0
cuda-runtime=12.1.0=0
cycler=0.12.1=pyhd8ed1ab_0
debugpy=1.8.0=py310hc6cd4ac_1
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py310h06a4308_0
fontconfig=2.14.1=hef1e5e3_0
freetype=2.10.4=hca18f0e_2
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h0b41bf4_3
glib=2.78.0=hfc55251_0
glib-tools=2.78.0=hfc55251_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
greenlet=3.0.1=py310hc6cd4ac_0
icu=70.1=h27087fc_0
idna=3.4=py310h06a4308_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
ipykernel=6.26.0=pyhf8b6a83_0
ipython=8.16.1=pyh0d859eb_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=py310h06a4308_0
joblib=1.2.0=py310h06a4308_0
jpeg=9e=h0b41bf4_3
jupyter_client=8.5.0=pyhd8ed1ab_0
jupyter_core=5.4.0=py310hff52083_0
kiwisolver=1.4.5=py310hd41b1e2_1
lame=3.100=h7b6447c_0
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=3.0=h295c915_0
libblas=3.9.0=19_linux64_openblas
libcblas=3.9.0=19_linux64_openblas
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.8.0.34=0
libcurand=10.3.4.52=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_2
libgfortran-ng=13.2.0=h69a702a_2
libgfortran5=13.2.0=ha4646dd_2
libglib=2.78.0=hebfc3b9_0
libgomp=13.2.0=h807b86a_2
libiconv=1.17=h166bdaf_0
libidn2=2.3.4=h5eee18b_0
liblapack=3.9.0=19_linux64_openblas
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libopenblas=0.3.24=pthreads_h413a1c8_0
libpng=1.6.39=h753d276_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.43.2=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_2
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=hd590300_0
libxcb=1.16=hd590300_0
libxml2=2.9.14=h22db469_4
libzlib=1.2.13=hd590300_5
markupsafe=2.1.1=py310h7f8727e_0
matplotlib-base=3.4.3=py310h23f4a51_2
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
ncurses=6.4=hcb278e6_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py310h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numexpr=2.8.7=py310h286c3b5_0
numpy=1.26.0=py310hb13e2d6_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.1.4=hd590300_0
packaging=23.2=pyhd8ed1ab_0
pandas=1.5.3=py310h1128e8f_0
parso=0.8.3=pyhd8ed1ab_0
pcre2=10.40=hc3806b6_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.0.1=py310ha6cbd5a_0
pip=23.3.1=pyhd8ed1ab_0
pixman=0.42.2=h59595ed_0
platformdirs=3.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.5=py310h2372a71_1
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycairo=1.23.0=py310hd1222b9_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.4.0=py310_torch_2.0.0_cu121
pygments=2.16.1=pyhd8ed1ab_0
pyopenssl=23.2.0=py310h06a4308_0
pyparsing=3.1.1=pyhd8ed1ab_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.13=hd12c33a_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.10=4_cp310
pytorch=2.0.1=cpu_py310hab5cca8_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cpu
pytz=2023.3.post1=pyhd8ed1ab_0
pyzmq=25.1.1=py310h795f18f_2
rdkit=2022.03.5=py310h1c297d8_0
readline=8.2=h8228510_1
reportlab=3.6.12=py310h5eee18b_0
requests=2.31.0=py310h06a4308_0
scikit-learn=1.3.0=py310h1128e8f_0
scipy=1.11.3=py310heeff2f4_0
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sqlalchemy=2.0.22=py310h2372a71_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=py310h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.13=h2797004_0
torchaudio=2.0.2=py310_cpu
torchvision=0.15.2=py310_cpu
tornado=6.3.3=py310h2372a71_1
tqdm=4.65.0=py310h2f386ee_0
traitlets=5.12.0=pyhd8ed1ab_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=1.26.18=py310h06a4308_0
wcwidth=0.2.8=pyhd8ed1ab_0
wheel=0.41.2=pyhd8ed1ab_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.4.2=h5eee18b_0
zeromq=4.3.5=h59595ed_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
